# docs-site/openapi/simpledeploy.yaml
openapi: 3.1.0
info:
  title: SimpleDeploy REST API
  version: 1.2.0
  summary: Single-binary Docker Compose deployer with proxy, backups, alerts, and metrics.
  description: |
    The SimpleDeploy HTTP API drives the dashboard UI and exposes every operation
    available in the CLI. All endpoints are mounted under the `/api` prefix.

    ## Authentication

    There are two ways to call the API:

    - **Session cookie** (`session`): set by `POST /auth/login` for browser clients.
      The cookie carries a JWT and is `HttpOnly`, `SameSite=Strict` (or `Lax` when
      TLS is off).
    - **API key bearer** (`Authorization: Bearer sd_...`): create at
      `POST /apikeys`. API keys belong to a user and inherit that user's role and
      per-app access grants.

    Most endpoints accept either auth method. A handful of dashboard-only routes
    require a real session.

    ## App scope

    Every app-scoped path is gated by an "app access" middleware. Super admins
    and admins can reach any app. Regular users only see apps explicitly granted
    to them via `POST /users/{id}/access`.

    ## Errors

    Errors are returned as `text/plain` with a one-line, generic message and the
    appropriate HTTP status. The detailed reason is logged server-side, never
    returned to the client. Treat any non-2xx status as a failure.
  license:
    name: MIT
    identifier: MIT
    url: https://github.com/vazra/simpledeploy/blob/main/LICENSE
  contact:
    name: SimpleDeploy on GitHub
...
servers:
...
security:
  - sessionAuth: []
  - apiKeyAuth: []
paths:
...
  /apps/{slug}/access:
    parameters: [{ $ref: '#/components/parameters/AppSlug' }]
    put:
      tags: [endpoints]
      summary: Update the IP allowlist for an app
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                allow:
                  type: string
                  description: Comma-separated IPs and/or CIDRs allowed to reach the app. Empty allows all traffic.
                  example: '1.2.3.4, 10.0.0.0/8'
      responses:
        '200':
          description: Access rules saved.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
                  allow: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/EditConflict' }
...
  /users/{id}/access:
    parameters: [{ $ref: '#/components/parameters/IdParam' }]
    post:
      tags: [users]
      summary: Grant a user access to an app
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [app_slug]
              properties:
                app_slug: { type: string }
      responses:
        '200':
          description: Granted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatusOk' }
  /users/{id}/access/{slug}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
      - $ref: '#/components/parameters/AppSlug'
    delete:
      tags: [users]
      summary: Revoke a user's access to an app
      responses:
        '200':
          description: Revoked.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatusOk' }
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session cookie set by `POST /auth/login`. Carries a JWT.
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: 'sd_<token>'
      description: Personal API key. Created via `POST /apikeys`.
  parameters:
    AppSlug:
      in: path
      name: slug
      required: true
      schema: { type: string }
      description: App slug (lowercase, dash-separated).
    IdParam:
      in: path
      name: id
      required: true
      schema: { type: integer, format: int64 }
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        text/plain:
          schema: { type: string }
...
    EditConflict:
      description: >-
        The change cannot be applied: a domain belongs to another app or the
        dashboard (text body), or the app's current compose file fails security
        checks (JSON body with every reason; fix and redeploy it first).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ViolationError' }
        text/plain:
          schema: { type: string }
  schemas:
    StatusOk:
      type: object
      properties:
        status: { type: string, example: ok }
...
    ViolationError:
      type: object
      properties:
        error: { type: string, example: compose file contains disallowed directives }
        violations:
          type: array
          items: { type: string }
          example: ['service "web": privileged mode not allowed']
...